лаба4.3

+++ laba11/src/main/java/my/snole/laba11/model/Point.java
@@ -1,62 +1,74 @@
 package my.snole.laba11.model;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    // Сложение точек (векторов)
+    /**
+     * Сложение векторов
+     */
     public Point add(Point other) {
         return new Point(this.x + other.x, this.y + other.y);
     }
 
-    // Вычитание точек (векторов)
+    /**
+     * Вычитание векторов
+     */
     public Point subtract(Point other) {
         return new Point(this.x - other.x, this.y - other.y);
     }
 
-    // Масштабирование точки (вектора)
+    /**
+     * Масштабирование вектора
+     */
     public Point scale(double scalar) {
-        return new Point((int)(this.x * scalar), (int)(this.y * scalar));
+        return new Point(this.x * scalar, this.y * scalar);
     }
 
-    // Нормализация вектора
+    /**
+     * Нормализация вектора
+     */
     public Point normalize() {
         double magnitude = this.getMagnitude();
         if (magnitude == 0) {
-            return new Point(0, 0); // или выбросить исключение
+            return new Point(0, 0);
         }
         return this.scale(1.0 / magnitude);
     }
 
-    // Получение длины вектора
+    /**
+     *  Получение длины вектора
+     */
     public double getMagnitude() {
-        return Math.sqrt(x*x + y*y);
+        return Math.sqrt(x * x + y * y);
     }
 
-    // Переопределение метода toString для удобства отладки
+    /**
+     * ! Метод для логов
+     */
     @Override
     public String toString() {
-        return String.format("Point(x=%d, y=%d)", x, y);
+        return String.format("Point(x=%.2f, y=%.2f)", x, y);
     }
 }
